requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
channels==4.3.1
channels_redis==4.3.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
factory_boy==3.3.3
Faker==37.8.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
kafka-python==2.2.15
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
